clean code + пагинация

+++ src/main/java/com/example/dto/GarageStats.java
@@ -1,4 +1,4 @@
 package com.example.dto;
 
-public record GarageStats(int count, double averageHp, String strongestModel) {
+public record GarageStats(long count, double averageHp, String strongestModel) {
 };
